retain reference links at the end of markdown files
Reference links are not emmitted as `Events` by the
`pulldown_cmark::Parser`, and the implementation of
`pulldown_cmark_to_cmark::State::finalize` only writes reference links
that were actually referenced somewhere in the markdown document.

To work around these limiataions we manually replace the links in the
`State` with the complete list of links from the `Parser`.